+++ sdk/src/client/reqwest/schema/mod.rs
@@ -12,7 +12,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use super::{DataType, PropertyDefinition, Schema, SchemaClient};
+pub(in crate) mod data;
+
+use crate::client::schema::{Schema, SchemaClient};
 
 use crate::client::reqwest::{fetch_entities_list, fetch_entity, post_batches};
 use crate::client::Client;
@@ -21,81 +23,6 @@ use crate::error::ClientError;
 use sawtooth_sdk::messages::batch::BatchList;
 
 const SCHEMA_ROUTE: &str = "schema";
-
-#[derive(Debug, Deserialize)]
-pub struct SchemaDto {
-    pub name: String,
-    pub description: String,
-    pub owner: String,
-    pub properties: Vec<PropertyDefinitionDto>,
-}
-
-impl From<&SchemaDto> for Schema {
-    fn from(d: &SchemaDto) -> Self {
-        Self {
-            name: d.name.to_string(),
-            description: d.description.to_string(),
-            owner: d.owner.to_string(),
-            properties: d.properties.iter().map(PropertyDefinition::from).collect(),
-        }
-    }
-}
-
-#[derive(Debug, Deserialize)]
-pub struct PropertyDefinitionDto {
-    pub name: String,
-    pub schema_name: String,
-    pub data_type: DataTypeDto,
-    pub required: bool,
-    pub description: String,
-    pub number_exponent: i64,
-    pub enum_options: Vec<String>,
-    pub struct_properties: Vec<PropertyDefinitionDto>,
-}
-
-impl From<&PropertyDefinitionDto> for PropertyDefinition {
-    fn from(d: &PropertyDefinitionDto) -> Self {
-        Self {
-            name: d.name.to_string(),
-            schema_name: d.schema_name.to_string(),
-            data_type: DataType::from(&d.data_type),
-            required: d.required,
-            description: d.description.to_string(),
-            number_exponent: d.number_exponent,
-            enum_options: d.enum_options.iter().map(String::from).collect(),
-            struct_properties: d
-                .struct_properties
-                .iter()
-                .map(PropertyDefinition::from)
-                .collect(),
-        }
-    }
-}
-
-#[derive(Deserialize, Debug)]
-pub enum DataTypeDto {
-    Bytes,
-    Boolean,
-    Number,
-    String,
-    Enum,
-    Struct,
-    LatLong,
-}
-
-impl From<&DataTypeDto> for DataType {
-    fn from(d: &DataTypeDto) -> Self {
-        match *d {
-            DataTypeDto::Bytes => DataType::Bytes,
-            DataTypeDto::Boolean => DataType::Boolean,
-            DataTypeDto::Number => DataType::Number,
-            DataTypeDto::String => DataType::String,
-            DataTypeDto::Enum => DataType::Enum,
-            DataTypeDto::Struct => DataType::Struct,
-            DataTypeDto::LatLong => DataType::LatLong,
-        }
-    }
-}
 
 /// The Reqwest implementation of the Schema client
 pub struct ReqwestSchemaClient {
@@ -134,8 +61,11 @@ impl SchemaClient for ReqwestSchemaClient {
     /// * `name` - the name of the schema (identifier)
     /// * `service_id` - optional - the service ID to fetch the schema from
     fn get_schema(&self, name: String, service_id: Option<&str>) -> Result<Schema, ClientError> {
-        let dto =
-            fetch_entity::<SchemaDto>(&self.url, format!("{}/{}", SCHEMA_ROUTE, name), service_id)?;
+        let dto = fetch_entity::<data::Schema>(
+            &self.url,
+            format!("{}/{}", SCHEMA_ROUTE, name),
+            service_id,
+        )?;
         Ok(Schema::from(&dto))
     }
 
@@ -145,7 +75,7 @@ impl SchemaClient for ReqwestSchemaClient {
     ///
     /// * `service_id` - optional - the service ID to fetch the schemas from
     fn list_schemas(&self, service_id: Option<&str>) -> Result<Vec<Schema>, ClientError> {
-        let dto_vec = fetch_entities_list::<SchemaDto>(
+        let dto_vec = fetch_entities_list::<data::Schema>(
             &self.url,
             SCHEMA_ROUTE.to_string(),
             service_id,
